improve map and fog

+++ AppleCinnamon/Options/SimplexOptions.cs
@@ -8,11 +8,11 @@ public readonly struct SimplexOptions
     public readonly int Octaves;
     public readonly double Frequency;
     public readonly double Amplitude;
-    public readonly int Offset;
+    public readonly double Offset;
     public readonly double Factor;
     public readonly Random Random;
 
-    public SimplexOptions(int octaves, double frequency, double amplitude, int offset, double factor, int seed)
+    public SimplexOptions(int octaves, double frequency, double amplitude, double offset, double factor, int seed)
     {
         Seed = seed;
         Octaves = octaves;
